requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==4.2.2
djangorestframework==3.15.1
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iniconfig==2.0.0
msgpack==1.0.8
outcome==1.3.0.post0
packaging==24.0
pluggy==1.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.7.0
pyOpenSSL==24.1.0
PySocks==1.7.1
pytest==8.1.1
python-dotenv==1.0.0
redis==5.0.7
requests==2.32.3
selenium==4.19.0
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.25.0
trio-websocket==0.11.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==2.2.1
websockets==12.0
wsproto==1.2.0
zope.event==5.0
zope.interface==6.2
